refactor Grid & Cell
