+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod configs;
-pub mod maze;
+pub mod grid;
 pub mod minigame;
 pub mod player;
 pub mod terrain;
